I have completed all the code modifications and verification tests according to the approved implementation plan. All 11 bugs and database integrations are now fully resolved.
Summary of Completed Tasks:
Repository Ingestion & Parser: Fixed

parser.py
 tree-sitter language loading crashes and byte-offset string slicing shifts.
Qdrant & Neo4j Pipelines: Replaced mock ingestion inside

worker.py
 with actual embedding generation (via litellm), Qdrant chunk storage, and Neo4j import-based dependency graph building.
Microservice Security & CORS: Restricted allowed origins in chat-service to allow secure cookie credentials and prevented Integer database overflows on User IDs.
OAuth Error Safety: Gracefully handled canceled login redirection instead of throwing FastAPI 422 errors.
Frontend Reliability: Solved ESLint type casting errors blocking compiler builds, resolved SSE chunk fragmentation drops using buffering, and added settings state persistence via localStorage.

+++ backend/auth-service/src/domain/models.py
@@ -1,6 +1,6 @@
 import uuid
 from datetime import datetime, timezone
-from sqlalchemy import Column, String, Integer, DateTime
+from sqlalchemy import Column, String, Integer, BigInteger, DateTime
 from sqlalchemy.dialects.postgresql import UUID
 from src.infrastructure.database import Base
 
@@ -8,7 +8,7 @@ class User(Base):
     __tablename__ = "users"
 
     id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4, index=True)
-    github_id = Column(Integer, unique=True, index=True, nullable=False)
+    github_id = Column(BigInteger, unique=True, index=True, nullable=False)
     email = Column(String, unique=True, index=True, nullable=False)
     name = Column(String, nullable=True)
     avatar_url = Column(String, nullable=True)
